+++ src/main/java/racingcar/domain/number/NumberGenerator.java
@@ -1,0 +1,5 @@
+package racingcar.domain.number;
+
+public interface NumberGenerator {
+    int generate();
+}
