+++ src/main/java/racingcar/domain/number/RandomNumberGenerator.java
@@ -1,17 +1,15 @@
-package racingcar.domain;
+package racingcar.domain.number;
 
 import java.util.Random;
 
-public class RandomNumber {
+public class RandomNumberGenerator implements NumberGenerator {
     private static final int MINIMUM_NUMBER = 0;
     private static final int MAXIMUM_NUMBER = 9;
 
     private static final Random random = new Random();
 
-    private RandomNumber() {
-    }
-
-    public static int make() {
+    @Override
+    public int generate() {
         return MINIMUM_NUMBER + random.nextInt(MAXIMUM_NUMBER - MINIMUM_NUMBER + 1);
     }
 }
